Implement authorization using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -62,6 +62,9 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if(\Gate::denies('update-question',$question)){
+            abort(403,'Access denied');
+        }
         return view('questions.edit',compact('question'));
     }
 
@@ -74,8 +77,11 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
-        $question->update($request->only('title','body'));
+        if(\Gate::denies('update-question',$question)){
+            abort(403,'Access denied');
+        }
 
+        $question->update($request->only('title','body'));
         return redirect()->route('questions.index')->with('success','Your question has been updated.');
     }
 
@@ -87,8 +93,11 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
-        $question->delete();
+        if(\Gate::denies('delete-question',$question)){
+            abort(403,'Access denied');
+        }
 
+        $question->delete();
         return redirect()->route('questions.index')->with('success','Your question has been deleted.');
     }
 }
